Update pol.html
Tried to shorten the intro paragraph to the most necessary bits and move other (important) points to the sub-sections. Also included UAAG as it is often relevant to internal policies and for some types of mobile apps. Also tried to abbreviate a lot of the text without losing the main points made. Hope I succeeded with some of that, at least to show you what I mean ;)