Possible changes to "What is this doc for" section
This section seems out-of-kilter to me.  Partly it's the switch from 3rd person to 2nd person, and adding to that is repeating the word "This" at the start of the first two sentences.  Also, the second half of the second sentence as best I can tell is actually saying what it is not for.  Anyways, I've made a stab at changes reflecting what I think it is you're trying to say.  Please feel free to disregard -- especially if I'm misinterpreting.  I'm not wanting to change the intended meaning.