First paragraph suggestions
This may be a repeat.  Apologies if so.  But it looks as if I forgot to make this into a pull request so trying again....  Paragraph 1 is so important, so giving it a bit of extra love....  (1) The first sentence is a bit hard to read and thus unclear, so I've tried to make it a bit smoother. (2) In the penultimate sentence there's a missing comma before however.  (3) In that same sentence, the words "an organizational or project" bog the sentence down and could easily distract the reader from the main point of the sentence, i.e. the immediate need to ID and remove barriers regardless of who or what.  Feel free to use whatever is helpful and disregard the rest.