Minor comma adjustments in "Next Steps"
I've noticed a number of places with unneeded commas, but have ignored them because they are in longish sentences.  However, here are comma is missing, so I've added it and gone ahead and deleted one on those unnecessary ones.

Also, I got lost at this point in the second sentence of the second paragraph: "reviewing company policies, such as procurement, to incorporate accessibility, or development infrastructure to ensure accessibility is baked in by default."  It needs work, but I'm unsure what you're trying to say at that point.  It's great up until that point.